Minor code style fixes, updated readme

+++ ostree.sh
@@ -1,6 +1,9 @@
-#!/usr/bin/env bash
+#!/bin/bash
 set -o errexit # Exit on non-zero status
 set -o nounset # Error on unset variables
+
+# Found unquoted vars with this regex:
+# (?<!")[$][{]OSTREE_(SYS|DEV|OPT|REP)_(BOOT|ROOT|TREE|HOME|DISK|NAME|KARG)(_LABEL)?[}](?!")
 
 # [ENVIRONMENT]: OVERRIDE DEFAULTS
 function ENV_CREATE_OPTS {
@@ -51,7 +54,7 @@ function ENV_VERIFY_LOCAL {
 function ENV_CREATE_DEPS {
 	# Skip in OSTree as filesystem is read-only
 	if ! ENV_VERIFY_LOCAL 2>/dev/null; then
-		pacman --noconfirm --sync --needed $@
+		pacman --noconfirm --sync --needed "$@"
 	fi
 }
 
@@ -71,22 +74,22 @@ function DISK_CREATE_LAYOUT {
 # [DISK]: FILESYSTEM (ESP+XFS)
 function DISK_CREATE_FORMAT {
 	ENV_CREATE_DEPS dosfstools xfsprogs
-	mkfs.vfat -n "${OSTREE_SYS_BOOT_LABEL}" -F 32 ${OSTREE_DEV_BOOT}
-	mkfs.xfs -L ${OSTREE_SYS_ROOT_LABEL} -f "${OSTREE_DEV_ROOT}" -n ftype=1
-	mkfs.xfs -L "${OSTREE_SYS_HOME_LABEL}" -f ${OSTREE_DEV_HOME} -n ftype=1
+	mkfs.vfat -n "${OSTREE_SYS_BOOT_LABEL}" -F 32 "${OSTREE_DEV_BOOT}"
+	mkfs.xfs -L "${OSTREE_SYS_ROOT_LABEL}" -f "${OSTREE_DEV_ROOT}" -n ftype=1
+	mkfs.xfs -L "${OSTREE_SYS_HOME_LABEL}" -f "${OSTREE_DEV_HOME}" -n ftype=1
 }
 
 # [DISK]: BUILD DIRECTORY
 function DISK_CREATE_MOUNTS {
-	mount --mkdir ${OSTREE_DEV_ROOT} ${OSTREE_SYS_ROOT}
-	mount --mkdir ${OSTREE_DEV_BOOT} ${OSTREE_SYS_ROOT}/boot/efi
+	mount --mkdir "${OSTREE_DEV_ROOT}" "${OSTREE_SYS_ROOT}"
+	mount --mkdir "${OSTREE_DEV_BOOT}" "${OSTREE_SYS_ROOT}"/boot/efi
 }
 
 # [OSTREE]: FIRST INITIALIZATION
 function OSTREE_CREATE_REPO {
 	ENV_CREATE_DEPS ostree which
-	ostree admin init-fs --sysroot="${OSTREE_SYS_ROOT}" --modern ${OSTREE_SYS_ROOT}
-	ostree admin stateroot-init --sysroot="${OSTREE_SYS_ROOT}" ${OSTREE_REP_NAME}
+	ostree admin init-fs --sysroot="${OSTREE_SYS_ROOT}" --modern "${OSTREE_SYS_ROOT}"
+	ostree admin stateroot-init --sysroot="${OSTREE_SYS_ROOT}" "${OSTREE_REP_NAME}"
 	ostree init --repo="${OSTREE_SYS_ROOT}/ostree/repo" --mode='bare'
 	ostree config --repo="${OSTREE_SYS_ROOT}/ostree/repo" set sysroot.bootprefix 1
 }
@@ -117,17 +120,17 @@ function OSTREE_CREATE_ROOTFS {
 	# Skip Podman layer cache if requested
 	if [[ ${PODMAN_OPT_NOCACHE} == 1 ]]; then
 		local PODMAN_OPT_BUILD=(
-			${PODMAN_OPT_BUILD[@]}
+			"${PODMAN_OPT_BUILD[@]}"
 			--no-cache='1'
 		)
 	fi
 
 	# Podman: create rootfs from multiple Containerfiles
 	for TARGET in ${PODMAN_OPT_BUILDFILE//,/ }; do
-		local PODMAN_OPT_IMG=(${TARGET%:*})
-		local PODMAN_OPT_TAG=(${TARGET#*:})
-		podman ${PODMAN_OPT_GLOBAL[@]} build \
-			${PODMAN_OPT_BUILD[@]} \
+		local PODMAN_OPT_IMG=("${TARGET%:*}")
+		local PODMAN_OPT_TAG=("${TARGET#*:}")
+		podman "${PODMAN_OPT_GLOBAL[@]}" build \
+			"${PODMAN_OPT_BUILD[@]}" \
 			--file="${PODMAN_OPT_IMG}" \
 			--tag="${PODMAN_OPT_TAG}" \
 			--cap-add='SYS_ADMIN' \
@@ -140,76 +143,76 @@ function OSTREE_CREATE_ROOTFS {
 	done
 
 	# Ostreeify: retrieve rootfs (workaround: `podman build --output local` doesn't preserve ownership)
-	rm -rf ${OSTREE_SYS_TREE}
-	mkdir -p ${OSTREE_SYS_TREE}
-	podman ${PODMAN_OPT_GLOBAL[@]} export $(podman ${PODMAN_OPT_GLOBAL[@]} create "${PODMAN_OPT_TAG}" bash) | tar -xC "${OSTREE_SYS_TREE}"
+	rm -rf "${OSTREE_SYS_TREE}"
+	mkdir -p "${OSTREE_SYS_TREE}"
+	podman "${PODMAN_OPT_GLOBAL[@]}" export $(podman "${PODMAN_OPT_GLOBAL[@]}" create "${PODMAN_OPT_TAG}" bash) | tar -xC "${OSTREE_SYS_TREE}"
 }
 
 # [OSTREE]: DIRECTORY STRUCTURE (https://ostree.readthedocs.io/en/stable/manual/adapting-existing)
 function OSTREE_CREATE_LAYOUT {
 	# Doing it here allows the container to be runnable/debuggable and Containerfile reusable
-	mv ${OSTREE_SYS_TREE}/etc ${OSTREE_SYS_TREE}/usr/
+	mv "${OSTREE_SYS_TREE}"/etc "${OSTREE_SYS_TREE}"/usr/
 
-	rm -r "${OSTREE_SYS_TREE}"/home
-	ln -s var/home ${OSTREE_SYS_TREE}/home
+	rm -r "${OSTREE_SYS_TREE:?}"/home
+	ln -s var/home "${OSTREE_SYS_TREE}"/home
 
-	rm -r "${OSTREE_SYS_TREE}"/mnt
-	ln -s var/mnt ${OSTREE_SYS_TREE}/mnt
+	rm -r "${OSTREE_SYS_TREE:?}"/mnt
+	ln -s var/mnt "${OSTREE_SYS_TREE}"/mnt
 
-	rm -r ${OSTREE_SYS_TREE}/opt
-	ln -s var/opt ${OSTREE_SYS_TREE}/opt
+	rm -r "${OSTREE_SYS_TREE:?}"/opt
+	ln -s var/opt "${OSTREE_SYS_TREE}"/opt
 
-	rm -r ${OSTREE_SYS_TREE}/root
-	ln -s var/roothome ${OSTREE_SYS_TREE}/root
+	rm -r "${OSTREE_SYS_TREE:?}"/root
+	ln -s var/roothome "${OSTREE_SYS_TREE}"/root
 
-	rm -r ${OSTREE_SYS_TREE}/srv
-	ln -s var/srv ${OSTREE_SYS_TREE}/srv
+	rm -r "${OSTREE_SYS_TREE:?}"/srv
+	ln -s var/srv "${OSTREE_SYS_TREE}"/srv
 
-	mkdir ${OSTREE_SYS_TREE}/sysroot
-	ln -s sysroot/ostree ${OSTREE_SYS_TREE}/ostree
+	mkdir "${OSTREE_SYS_TREE:?}"/sysroot
+	ln -s sysroot/ostree "${OSTREE_SYS_TREE}"/ostree
 
-	rm -r "${OSTREE_SYS_TREE}"/usr/local
-	ln -s ../var/usrlocal ${OSTREE_SYS_TREE}/usr/local
+	rm -r "${OSTREE_SYS_TREE:?}"/usr/local
+	ln -s ../var/usrlocal "${OSTREE_SYS_TREE}"/usr/local
 
 	printf >&1 '%s\n' 'Creating tmpfiles'
-	echo 'd /var/home 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/lib 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/log/journal 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/mnt 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/opt 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/home 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/lib 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/log/journal 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/mnt 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/opt 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
 	echo 'd /var/roothome 0700 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/srv 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/usrlocal 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/usrlocal/bin 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/usrlocal/etc 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/usrlocal/games 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/usrlocal/include 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/usrlocal/lib 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/usrlocal/man 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/usrlocal/sbin 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/usrlocal/share 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /var/usrlocal/src 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
-	echo 'd /run/media 0755 root root -' >> ${OSTREE_SYS_TREE}/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/srv 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/usrlocal 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/usrlocal/bin 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/usrlocal/etc 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/usrlocal/games 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/usrlocal/include 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/usrlocal/lib 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/usrlocal/man 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/usrlocal/sbin 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/usrlocal/share 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /var/usrlocal/src 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
+	echo 'd /run/media 0755 root root -' >> "${OSTREE_SYS_TREE}"/usr/lib/tmpfiles.d/ostree-0-integration.conf
 
 	# Only retain information about Pacman packages in new rootfs
-	mv ${OSTREE_SYS_TREE}/var/lib/pacman ${OSTREE_SYS_TREE}/usr/lib/
+	mv "${OSTREE_SYS_TREE}"/var/lib/pacman "${OSTREE_SYS_TREE}"/usr/lib/
 	sed -i \
 		-e 's|^#\(DBPath\s*=\s*\).*|\1/usr/lib/pacman|g' \
 		-e 's|^#\(IgnoreGroup\s*=\s*\).*|\1modified|g' \
-		${OSTREE_SYS_TREE}/usr/etc/pacman.conf
+		"${OSTREE_SYS_TREE}"/usr/etc/pacman.conf
 
 	# Allow Pacman to store update notice id during unlock mode
-	mkdir ${OSTREE_SYS_TREE}/usr/lib/pacmanlocal
+	mkdir "${OSTREE_SYS_TREE}"/usr/lib/pacmanlocal
 
-	# OSTree mounts /ostree/deploy/${OSTREE_REP_NAME}/var to /var
-	rm -r "${OSTREE_SYS_TREE}"/var/*
+	# OSTree mounts /ostree/deploy/"${OSTREE_REP_NAME}"/var to /var
+	rm -r "${OSTREE_SYS_TREE:?}"/var/*
 }
 
 # [OSTREE]: CREATE COMMIT
 function OSTREE_DEPLOY_IMAGE {
 	# Update repository and boot entries in GRUB2
 	ostree commit --repo="${OSTREE_SYS_ROOT}/ostree/repo" --branch="${OSTREE_REP_NAME}/latest" --tree=dir="${OSTREE_SYS_TREE}"
-	ostree admin deploy --sysroot="${OSTREE_SYS_ROOT}" --karg="root=LABEL=${OSTREE_SYS_ROOT_LABEL} rw ${OSTREE_SYS_KARG}" --os="${OSTREE_REP_NAME}" ${OSTREE_OPT_NOMERGE} --retain ${OSTREE_REP_NAME}/latest
+	ostree admin deploy --sysroot="${OSTREE_SYS_ROOT}" --karg="root=LABEL=${OSTREE_SYS_ROOT_LABEL} rw ${OSTREE_SYS_KARG}" --os="${OSTREE_REP_NAME}" "${OSTREE_OPT_NOMERGE}" --retain "${OSTREE_REP_NAME}"/latest
 }
 
 # [OSTREE]: UNDO COMMIT
@@ -220,18 +223,19 @@ function OSTREE_REVERT_IMAGE {
 # [BOOTLOADER]: FIRST BOOT
 # | Todo: improve grub-mkconfig
 function BOOTLOADER_CREATE {
-	grub-install --target='x86_64-efi' --efi-directory="${OSTREE_SYS_ROOT}/boot/efi" --boot-directory="${OSTREE_SYS_ROOT}/boot/efi/EFI" --bootloader-id="${OSTREE_REP_NAME}" --removable ${OSTREE_DEV_BOOT}
+	grub-install --target='x86_64-efi' --efi-directory="${OSTREE_SYS_ROOT}/boot/efi" --boot-directory="${OSTREE_SYS_ROOT}/boot/efi/EFI" --bootloader-id="${OSTREE_REP_NAME}" --removable "${OSTREE_DEV_BOOT}"
 
-	local OSTREE_SYS_PATH=$(ls -d "${OSTREE_SYS_ROOT}"/ostree/deploy/"${OSTREE_REP_NAME}"/deploy/* | head -n 1)
+	local OSTREE_SYS_PATH
+	OSTREE_SYS_PATH=$(ls -d "${OSTREE_SYS_ROOT}"/ostree/deploy/"${OSTREE_REP_NAME}"/deploy/* | head -n 1)
 
-	rm -rfv "${OSTREE_SYS_PATH}"/boot/*
-	mount --mkdir --rbind "${OSTREE_SYS_ROOT}"/boot ${OSTREE_SYS_PATH}/boot
-	mount --mkdir --rbind "${OSTREE_SYS_ROOT}"/ostree ${OSTREE_SYS_PATH}/sysroot/ostree
+	rm -rfv "${OSTREE_SYS_PATH:?}"/boot/*
+	mount --mkdir --rbind "${OSTREE_SYS_ROOT}"/boot "${OSTREE_SYS_PATH}"/boot
+	mount --mkdir --rbind "${OSTREE_SYS_ROOT}"/ostree "${OSTREE_SYS_PATH}"/sysroot/ostree
 
-	for i in /dev /proc /sys; do mount -o bind $i ${OSTREE_SYS_PATH}${i}; done
-	chroot ${OSTREE_SYS_PATH} /bin/bash -c 'grub-mkconfig -o /boot/efi/EFI/grub/grub.cfg'
+	for i in /dev /proc /sys; do mount -o bind $i "${OSTREE_SYS_PATH}"${i}; done
+	chroot "${OSTREE_SYS_PATH}" /bin/bash -c 'grub-mkconfig -o /boot/efi/EFI/grub/grub.cfg'
 
-	umount --recursive ${OSTREE_SYS_ROOT}
+	umount --recursive "${OSTREE_SYS_ROOT}"
 }
 
 # [CLI]: TASK FINECONTROL
@@ -362,7 +366,7 @@ function CLI_SETUP {
 				;;
 
 				*)
-					if [[ $(type -t ${CLI_VAL}) == 'function' ]]; then
+					if [[ $(type -t "${CLI_VAL}") == 'function' ]]; then
 						ENV_CREATE_OPTS
 						${CLI_VAL}
 					fi
